requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-qdrant
qdrant-client
pypdf
python-docx
pandas
sentence-transformers
tqdm
torch
psycopg2-binary
sqlalchemy
python-dotenv
